requirements.txt
Django==1.6.5
PyYAML==3.11
cx-Oracle==5.1.3
django-tastypie==0.11.1
python-dateutil==2.2
python-mimeparse==0.1.4
simplejson==3.5.3
six==1.7.3
wsgiref==0.1.2
